+++ src/nginx_module/access_phase_handler.h
@@ -1,0 +1,3 @@
+#include <ngx_http.h>
+
+ngx_int_t FiretailAccessPhaseHandler(ngx_http_request_t *r);
