+++ src/nginx_module/config
@@ -3,7 +3,7 @@ ngx_addon_name=ngx_firetail_module
 FIRETAIL_SRCS="                                                             \
         $ngx_addon_dir/firetail_module.c                                    \
         $ngx_addon_dir/filter_context.c                                     \
-        $ngx_addon_dir/filter_firetail_send.c                               \
+        $ngx_addon_dir/access_phase_handler.c                               \
         $ngx_addon_dir/filter_response_body.c                               \
         $ngx_addon_dir/filter_headers.c                                     \
         "
@@ -11,7 +11,7 @@ FIRETAIL_SRCS="                                                             \
 FIRETAIL_DEPS="                                                             \
         $ngx_addon_dir/firetail_module.h                                    \
         $ngx_addon_dir/filter_context.h                                     \
-        $ngx_addon_dir/filter_firetail_send.h                               \
+        $ngx_addon_dir/access_phase_handler.h                               \
         $ngx_addon_dir/filter_response_body.h                               \
         $ngx_addon_dir/filter_headers.h                                     \
         "
