+++ src/nginx_module/filter_response_body.h
@@ -5,4 +5,5 @@
 #include <ngx_http.h>
 
 ngx_int_t FiretailResponseBodyFilter(ngx_http_request_t *request, ngx_chain_t *chain_head);
+
 #endif
