+++ src/nginx_module/firetail_module.h
@@ -3,24 +3,6 @@
 
 #include <ngx_http.h>
 
-typedef int (*CreateMiddlewareFunc)(void*, int);
-
-struct ValidateResponseBody_return {
-  int r0;
-  char* r1;
-};
-typedef struct ValidateResponseBody_return (*ValidateResponseBody)(char*, int, char*, int, char*, int, char*, int,
-                                                                   char*, int, void*, int, char*, int, void*, int, int,
-                                                                   void*, int);
-
-struct ValidateRequestBody_return {
-  int r0;
-  char* r1;
-};
-typedef struct ValidateRequestBody_return (*ValidateRequestBody)(void*, int, void*, int, void*, int, void*, int, void*,
-                                                                 int);
-
-// This config struct will hold our API key
 typedef struct {
   ngx_str_t FiretailApiToken;  // TODO: this should probably be a *ngx_str_t
   ngx_str_t FiretailUrl;
